Exception refactoring
