+++ src/main/java/svydovets/core/context/injector/AbstractInjector.java
@@ -2,7 +2,7 @@ package svydovets.core.context.injector;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import svydovets.exception.AutowireBeanException;
+import svydovets.core.exception.AutowireBeanException;
 import svydovets.exception.BeanCreationException;
 import svydovets.exception.FieldValueIllegalAccessException;
 
@@ -58,8 +58,8 @@ public abstract class AbstractInjector implements Injector {
             log.error(ERROR_AUTOWIRED_BEAN_EXCEPTION_MESSAGE);
             log.error(exception.getMessage());
 
-            throw new AutowireBeanException(String
-                    .format(ERROR_AUTOWIRED_BEAN_EXCEPTION_MESSAGE, fieldForInjection.getName()));
+            throw new AutowireBeanException(
+                String.format(ERROR_AUTOWIRED_BEAN_EXCEPTION_MESSAGE, fieldForInjection.getName()), exception);
         }
     }
