+++ src/main/java/svydovets/core/exception/AutowireBeanException.java
@@ -1,4 +1,4 @@
-package svydovets.exception;
+package svydovets.core.exception;
 
 public class AutowireBeanException extends RuntimeException{
     public AutowireBeanException(String message) {
